split out tree-finder-mockcontents into separate pkg
