+++ examples/simple/package.json
@@ -18,7 +18,8 @@
   "keywords": [],
   "license": "BSD-3-Clause",
   "dependencies": {
-    "tree-finder": "^0.0.7"
+    "tree-finder": "^0.0.7",
+    "tree-finder-mockcontents": "^0.0.1"
   },
   "devDependencies": {
     "html-webpack-plugin": "^5.0.0",
