+++ examples/simple/src/index.ts
@@ -6,9 +6,11 @@
  */
 // import "tree-finder";
 import {
-  mockContent,
   TreeFinderPanelElement,
 } from "tree-finder";
+import {
+  mockContent,
+} from "tree-finder-mockcontents";
 
 import "tree-finder/style/theme/material.css";
 import "../style";
