+++ packages/tree-finder-mockcontents/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "tree-finder-mockcontents",
+  "version": "0.0.1",
+  "description": "browser/tree widget implemented as a custom DOM element. Built on top of regular-table",
+  "homepage": "https://github.com/telamonian/tree-finder-mockcontents",
+  "bugs": {
+    "url": "https://github.com/telamonian/tree-finder-mockcontents/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/telamonian/tree-finder-mockcontents.git"
+  },
+  "license": "BSD-3-Clause",
+  "author": "Max Klein <telamonian@users.noreply.github.com>",
+  "files": [
+    "dist/**/*",
+    "lib/**/*.{d.ts,js,map}",
+    "style/**/*.{css,eot,gif,html,jpg,json,less,png,scss,svg,ttf,woff2}"
+  ],
+  "main": "lib/index.js",
+  "sideEffects": [
+    "style/**/*"
+  ],
+  "types": "lib/index.d.ts",
+  "scripts": {
+    "build": "tsc --build",
+    "build:watch": "tsc --build --watch",
+    "clean": "rimraf dist lib *.tsbuildinfo",
+    "clean:slate": "yarn run clean && rimraf node_modules",
+    "postpack": "shx rm README.md LICENSE",
+    "prepack": "shx cp ../../README.md . && shx cp ../../LICENSE .",
+    "prepublishOnly": "npm run clean && npm run build && npm run webpack:prod"
+  },
+  "dependencies": {
+    "faker": "^5.4.0",
+    "tree-finder": "^0.0.7"
+  },
+  "devDependencies": {
+    "@types/faker": "^5.1.7",
+    "typescript": "^4.1.3"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
