+++ packages/tree-finder-mockcontents/src/mockcontent.ts
@@ -4,37 +4,12 @@
  * This file is part of the tree-finder library, distributed under the terms of
  * the BSD 3 Clause license. The full license can be found in the LICENSE file.
  */
-import { Path, IContentRow } from "./content";
-import { Random } from "./util";
+// import * as faker from "faker";
 
-const DIR_NAMES = [
-  "able",
-  "baker",
-  "charlie",
-  "dog",
-  "easy",
-  "fox",
-  "george",
-  "how",
-  "item",
-  "jig",
-  "king",
-  "love",
-  "mike",
-  "nan",
-  "oboe",
-  "peter",
-  "queen",
-  "roger",
-  "sugar",
-  "tare",
-  "uncle",
-  "victor",
-  "william",
-  "xray",
-  "yoke",
-  "zebra",
-];
+import { Path, IContentRow } from "tree-finder";
+
+import { ALLIED_PHONETIC, Random } from "./util";
+
 
 interface IMockContentRow extends IContentRow {
   modified: Date;
@@ -60,7 +35,7 @@ export function mockContent(props: {path: Path, kind: string, modDays?: number, 
       writable,
       getChildren: async () => {
         const children = [];
-        const dirNames = randomize ? Random.shuffle(DIR_NAMES) : DIR_NAMES;
+        const dirNames = randomize ? Random.shuffle(ALLIED_PHONETIC) : ALLIED_PHONETIC;
 
         for (let i = 0; i < nchildren; i++) {
           children.push(mockContent({
